New forcings lists nanagment

+++ src/TurtleClass.py
@@ -18,7 +18,7 @@ def define_Turtle_Class(fieldset, param):
         deg = Variable('deg', to_write=False, dtype=np.float32, initial=111195)#1degree = 111,195 km approx
         alpha = Variable('alpha', to_write=False, dtype=np.float32)
         mode = Variable('mode', to_write=False, dtype=np.float32)
-        tstep = Variable('tstep', to_write=False, dtype=np.float32)
+        tstep = Variable('tstep', to_write=False, dtype=np.float64)
         #Sampling
         if param['key_alltracers']:
             T = Variable('T', initial=fieldset.T)
